+++ examples/data-sources/okta_label/datasource.tf
@@ -1,0 +1,11 @@
+resource "okta_label" "test" {
+  name = "test-label-replace_with_uuid"
+
+  values {
+    name = "test-value"
+  }
+}
+
+data "okta_label" "test" {
+  id = okta_label.test.id
+}