+++ examples/resources/okta_oauth2_v1_clients_role_api_access_management_admin/resource.tf
@@ -1,0 +1,4 @@
+resource "okta_oauth2_v1_clients_role_api_access_management_admin" "example" {
+  client_id = "<client_id>"
+  type      = "API_ACCESS_MANAGEMENT_ADMIN"
+}