+++ examples/resources/okta_template_sms/resource.tf
@@ -1,13 +1,9 @@
 resource "okta_template_sms" "example" {
+  name     = "My SMS Template"
   type     = "SMS_VERIFY_CODE"
   template = "Your $${org.name} code is: $${code}"
-  translations {
-    language = "en"
-    template = "Your $${org.name} code is: $${code}"
-  }
-
-  translations {
-    language = "es"
-    template = "Tu código de $${org.name} es: $${code}."
+  translations = {
+    "en" = "Your $${org.name} code is: $${code}"
+    "es" = "Tu código de $${org.name} es: $${code}."
   }
 }