+++ examples/resources/okta_threat_insight_settings/basic_updated.tf
@@ -8,5 +8,5 @@ resource "okta_network_zone" "test" {
 
 resource "okta_threat_insight_settings" "test" {
   action           = "block"
-  network_excludes = [okta_network_zone.test.id]
+  exclude_zones = [okta_network_zone.test.id]
 }