+++ examples/resources/okta_trusted_origin/okta_trusted_origin_updated.tf
@@ -1,6 +1,11 @@
 resource "okta_trusted_origin" "testAcc_replace_with_uuid" {
   name   = "testAcc-replace_with_uuid"
-  active = false
+  status = "INACTIVE"
   origin = "https://example2-replace_with_uuid.com"
-  scopes = ["CORS", "REDIRECT"]
+  scopes {
+    type = "CORS"
+  }
+  scopes {
+    type = "REDIRECT"
+  }
 }