+++ okta/resources/resources.go
@@ -53,6 +53,7 @@ const (
 	OktaIDaaSBrand                                    = "okta_brand"
 	OktaIDaaSBrands                                   = "okta_brands"
 	OktaIDaaSCaptcha                                  = "okta_captcha"
+	OktaIDaaSCaptchaOrgWide                           = "okta_org_captcha"
 	OktaIDaaSCaptchaOrgWideSettings                   = "okta_captcha_org_wide_settings"
 	OktaIDaaSDefaultPolicy                            = "okta_default_policy"
 	OktaIDaaSDomain                                   = "okta_domain"
@@ -175,4 +176,6 @@ const (
 	OktaIDaaSAuthenticatorWebauthnCustomAAGUIDs       = "okta_authenticator_webauthn_custom_aaguids"
 	OktaIDaaSAuthenticatorMethodWebauthn              = "okta_authenticator_method_webauthn"
 	OktaIDaaSPoliciesRuleSignOn                       = "okta_policies_rule_sign_on"
+	OktaGovernanceResourceOwner                       = "okta_resource_owner"
+	OktaGovernanceResourceOwnersCatalogResource       = "okta_resource_owners_catalog_resource"
 )