+++ okta/services/idaas/idaas.go
@@ -86,6 +86,7 @@ func providerIsClassicOrg(ctx context.Context, m interface{}) bool {
 
 func FWProviderResources() []func() resource.Resource {
 	rawResources := []func() resource.Resource{
+		newTrustedOriginResource,
 		newAppAccessPolicyAssignmentResource,
 		newAppOAuthRoleAssignmentResource,
 		newTrustedServerResource,
@@ -102,6 +103,7 @@ func FWProviderResources() []func() resource.Resource {
 		newGroupOwnersResource,
 		newAppSignOnPolicyResource,
 		newEmailTemplateSettingsResource,
+		newTemplateSmsResource,
 		newFeaturesResource,
 		newRealmResource,
 		newRealmAssignmentResource,
@@ -133,6 +135,24 @@ func FWProviderResources() []func() resource.Resource {
 		newIdentitySourceGroupMembershipResource,
 		newIdentitySourceImportResource,
 		newIdentitySourceUserResource,
+		newCaptchaResource,
+		newOrgCaptchaResource,
+		newOauth2V1ClientsRoleAccessCertificationsAdminResource,
+		newOauth2V1ClientsRoleAccessRequestsAdminResource,
+		newOauth2V1ClientsRoleApiAccessManagementAdminResource,
+		newOauth2V1ClientsRoleAppAdminResource,
+		newOauth2V1ClientsRoleCustomResource,
+		newOauth2V1ClientsRoleGroupMembershipAdminResource,
+		newOauth2V1ClientsRoleHelpDeskAdminResource,
+		newOauth2V1ClientsRoleOrgAdminResource,
+		newOauth2V1ClientsRoleReadOnlyAdminResource,
+		newOauth2V1ClientsRoleReportAdminResource,
+		newOauth2V1ClientsRoleSuperAdminResource,
+		newOauth2V1ClientsRoleUserAdminResource,
+		newOauth2V1ClientsRoleWorkflowsAdminResource,
+		newRoleSubscriptionResource,
+		newUserSubscriptionResource,
+		newThreatInsightSettingsResource,
 	}
 	// Wrap all resources with SafeResource for panic recovery
 	return resources.WrapResources(rawResources)
@@ -184,6 +204,11 @@ func FWProviderDataSources() []func() datasource.DataSource {
 		newAuthorizationServersPoliciesRuleDataSource,
 		newIamAssigneesUserDataSource,
 		newIamResourceSetDataSource,
+		newRoleSubscriptionDataSource,
+		newUserSubscriptionDataSource,
+		newThreatInsightSettingsDataSource,
+		newCaptchaDataSource,
+		newOrgCaptchaDataSource,
 	}
 }
 
@@ -221,7 +246,6 @@ func ProviderResources() map[string]*schema.Resource {
 		resources.OktaIDaaSAuthServerPolicyRule:          resourceAuthServerPolicyRule(),
 		resources.OktaIDaaSAuthServerScope:               resourceAuthServerScope(),
 		resources.OktaIDaaSBehavior:                      resourceBehavior(),
-		resources.OktaIDaaSCaptcha:                       resourceCaptcha(),
 		resources.OktaIDaaSCaptchaOrgWideSettings:        resourceCaptchaOrgWideSettings(),
 		resources.OktaIDaaSDomain:                        resourceDomain(),
 		resources.OktaIDaaSDomainCertificate:             resourceDomainCertificate(),
@@ -266,12 +290,8 @@ func ProviderResources() map[string]*schema.Resource {
 		resources.OktaIDaaSProfileMapping:                resourceProfileMapping(),
 		// resources.OktaIDaaSRateLimiting:                  resourceRateLimiting(),
 		resources.OktaIDaaSResourceSet:                resourceResourceSet(),
-		resources.OktaIDaaSRoleSubscription:           resourceRoleSubscription(),
 		resources.OktaIDaaSSecurityNotificationEmails: resourceSecurityNotificationEmails(),
-		resources.OktaIDaaSTemplateSms:                resourceTemplateSms(),
 		resources.OktaIDaaSTheme:                      resourceTheme(),
-		resources.OktaIDaaSThreatInsightSettings:      resourceThreatInsightSettings(),
-		resources.OktaIDaaSTrustedOrigin:              resourceTrustedOrigin(),
 		resources.OktaIDaaSUser:                       resourceUser(),
 		resources.OktaIDaaSUserAdminRoles:             resourceUserAdminRoles(),
 		resources.OktaIDaaSUserBaseSchemaProperty:     resourceUserBaseSchemaProperty(),
@@ -320,7 +340,6 @@ func ProviderDataSources() map[string]*schema.Resource {
 		resources.OktaIDaaSNetworkZone:              dataSourceNetworkZone(),
 		resources.OktaIDaaSPolicy:                   dataSourcePolicy(),
 		resources.OktaIDaaSPolicyRulePassword:       dataSourcePolicyRulePassword(),
-		resources.OktaIDaaSRoleSubscription:         dataSourceRoleSubscription(),
 		resources.OktaIDaaSTheme:                    dataSourceTheme(),
 		resources.OktaIDaaSThemes:                   dataSourceThemes(),
 		resources.OktaIDaaSTrustedOrigins:           dataSourceTrustedOrigins(),